intoduce contexts.py, exception.py, git_operations.py
