+++ git_machete/git_operations.py
@@ -1,0 +1,5 @@
+
+
+class GitContext:
+    def __init__(self):
+        pass
